moved index name calculation to new class and added test
